refactored: Dialogs
